Fragmenting template.js
