+++ index.html
@@ -14,7 +14,7 @@
     <link rel="icon" type="image/png" sizes="16x16" href="assets/images/favicon_package/favicon-16x16.png">
     <link rel="manifest" href="assets/images/favicon_package/site.webmanifest">
     <link rel="mask-icon" href="assets/images/favicon_package/safari-pinned-tab.svg" color="#5bbad5">
-
+    <script type="module"  src="scripts/main.js"></script>
     <link rel="stylesheet" href="styles/styles.css">
     <title>Placeholder</title>
 </head>
@@ -354,6 +354,4 @@
     <output class="loader"></output>
 </div>
 <!-- SPINNER -->
-  
 
-<script type="module"  src="scripts/main.js"></script>
